Refactor commands into modular handlers
